added pokemon up to Magneton.
Tried upgrading in place the game version so that linux would stop bugging me, but now it's a hex string, so I'm not pushing the GetThemPokes.uproject, who know what that would break if sent upstream.
Now officially over halfway through the first pass of the gen one pokemon, still need to add in the pokemon abilities following the example set in 001's moveList.