BigListOPokemon is now up to number 121 Starmie.
Note: Apparently pokemondb shows pokemon types that do not exist in gen 1, so a second pass should be done to replace the illicit pokemon types as well as add in the expected exp on defeat. 
Adding the list of moves each pokemon learns will be relegated to the third pass now.